ptc_mfb2pcie_axi [FEATURE]: the design adjusted to be more general
The design was remade to configurate the output according to the number
of regions given by the generic parameter and not by the DEVICE generic.
That is because the UltraScale devices can also have the disabled
Straddling option on the 256b variant of the bus. The AXI_DATA_WIDTH
generic parameter has also been removed because some values of this
generic can clash with MFB parameters being set so the latter has been
kept because it allows for more fine grained configuration.

The CLK and RESET ports have been removed because the component is not
synchronous and the RESET was not used at all. Those ports and above
mentioned generics have also been removed from this component
instantiation in the PTC module.