tx_dma_calypte [FEATURE]: change internal architecture to support out-of-order mem writes
The previous design did not support the reception on the adresses that
were not sorted from lower to higher. The FIFOs as data buffers had to
be therefore removed and were replaced by usual BRAMs. The pointer logic
have returned and the DMAs now work as inverse RX DMA Calypte. The whole
communication change was also changed so minimal footprint on chip is
taken. The splitting to different DMA channels is done only by data buffers.