rx_dma_calypte [FEATURE]: add support for Intel devices
Those devices are distinguished from the Xilinx that the PCIe header is
not prepended before data but send in a special meta signal. The major
change is therefore in the hdr_insertor component because the shift of
data changes depending on if there needs to be a PCIe header prepended
or not.