mfb_generator [FEATURE]: add support for burst mode
The activation of this mode causes the generator to stop transmission
after a burst_size amount of packets has been sent. In order to send a
new burst, the generator needs to be disabled and enabled again in this cycles.

This new mode is, however, not supported by the CORE_PACP architecture
so when using that, the burst mode will be automatically disabled.