data_logger.py [BUGFIX]: fix slicing when reading CTRLO_WIDTH
The reading of slices of CTRLO signal has been set in such a way that the
0th slice (slices occur when the width of the CTRLO signal is wider than
the set width of the MI bus) has been skipped. Therefore, the lower bits
could not be set. The correction has been made that only when the
current slice has already been written, does not need to be selected
again.